# frontend/src/api/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "ChatContext": {
        "properties": {
          "active_symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Symbol"
          },
          "selection": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SelectionContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "selections": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SelectionContext"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selections"
          },
          "source_tab": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 1,
                "pattern": "^[^\\x00-\\x1F\\x7F]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Tab"
          },
          "source_view": {
            "anyOf": [
              {
                "enum": [
                  "dashboard",
                  "command_palette"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source View"
          },
          "view": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "View"
          }
        },
        "title": "ChatContext",
        "type": "object"
      },
      "ChatMessage": {
        "properties": {
          "content": {
            "description": "message content",
            "title": "Content",
            "type": "string"
          },
          "role": {
            "description": "message role",
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role",
          "content"
        ],
        "title": "ChatMessage",
        "type": "object"
      },
      "ChatOptions": {
        "properties": {
          "locale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locale"
          },
          "output_mode": {
            "anyOf": [
              {
                "enum": [
                  "chat",
                  "brief",
                  "investment_report"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Mode"
          },
          "strict_selection": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strict Selection"
          },
          "trace_raw_override": {
            "anyOf": [
              {
                "enum": [
                  "on",
                  "off",
                  "inherit"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Raw Override"
          }
        },
        "title": "ChatOptions",
        "type": "object"
      },
...
      "ExecuteRequest": {
        "description": "Chat、Dashboard handoff 与报告生成共用的执行请求。",
        "properties": {
          "analysis_depth": {
            "anyOf": [
              {
                "enum": [
                  "quick",
                  "report",
                  "deep_research"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Depth"
          },
          "budget": {
            "anyOf": [
              {
                "maximum": 10.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget"
          },
          "context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChatContext"
              },
              {
                "type": "null"
              }
            ],
            "description": "Ephemeral UI context"
          },
          "history": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ChatMessage"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Visible conversation history",
            "title": "History"
          },
          "options": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChatOptions"
              },
              {
                "type": "null"
              }
            ],
            "description": "Chat execution options"
          },
          "output_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Mode"
          },
          "query": {
            "description": "Analysis query",
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional correlation ID",
            "title": "Run Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Conversation session ID",
            "title": "Session Id"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "tickers": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tickers"
          },
          "trace_raw": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Raw"
          }
        },
        "required": [
          "query"
        ],
        "title": "ExecuteRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SelectionContext": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "snippet": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snippet"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "ts": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ts"
          },
          "type": {
            "enum": [
              "news",
              "filing",
              "doc",
              "report",
              "risk",
              "insight",
              "url",
              "web",
              "article"
            ],
            "title": "Type",
            "type": "string"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "type",
          "id",
          "title"
        ],
        "title": "SelectionContext",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "可信行情、AI Prediction 与证据化研究 API",
    "title": "FinSight API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/execute": {
      "post": {
        "operationId": "execute_endpoint_api_execute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Execute Endpoint",
        "tags": [
          "Execution"
        ]
      }
    },
    "/api/execute/runs/{run_id}/cancel": {
      "post": {
        "operationId": "cancel_run_api_execute_runs__run_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Run",
        "tags": [
          "Execution"
        ]
      }
    },
    "/api/execute/runs/{run_id}/events": {
      "get": {
        "operationId": "replay_events_api_execute_runs__run_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after_seq",
            "required": false,
            "schema": {
              "default": 0,
              "title": "After Seq",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replay Events",
        "tags": [
          "Execution"
        ]
      }
    },
...
  }
}